+++ src/components/dialogs/InviteDialog/EditAndSendInviteDialog.vue
@@ -35,7 +35,7 @@
     import { getCurrentInstance, inject } from 'vue';
     import { useI18n } from 'vue-i18n-bridge';
     import { instanceRequest, inviteMessagesRequest, notificationRequest } from '../../../api';
-    import utils from '../../../classes/utils';
+    import { parseLocation } from '../../../composables/instance/utils';
 
     const { t } = useI18n();
     const instance = getCurrentInstance();
@@ -104,7 +104,7 @@
                     const receiverUserId = J.userIds.shift();
                     if (receiverUserId === API.currentUser.id) {
                         // can't invite self!?
-                        const L = utils.parseLocation(J.worldId);
+                        const L = parseLocation(J.worldId);
                         instanceRequest
                             .selfInvite({
                                 instanceId: L.instanceId,
