+++ src/components/dialogs/PreviousInstancesDialog/PreviousInstancesWorldDialog.vue
@@ -68,6 +68,7 @@
 
 <script>
     import utils from '../../../classes/utils';
+    import { parseLocation } from '../../../composables/instance/utils';
     import database from '../../../service/database';
 
     export default {
@@ -138,7 +139,7 @@
                 database.getpreviousInstancesByWorldId(D.worldRef).then((data) => {
                     const array = [];
                     for (const ref of data.values()) {
-                        ref.$location = utils.parseLocation(ref.location);
+                        ref.$location = parseLocation(ref.location);
                         if (ref.time > 0) {
                             ref.timer = utils.timeToText(ref.time);
                         } else {
