+++ src/composables/avatar/utils.js
@@ -30,4 +30,96 @@ function parseAvatarUrl(avatar) {
     // return void 0;
 }
 
-export { storeAvatarImage, parseAvatarUrl };
+function getPlatformInfo(unityPackages) {
+    var pc = {};
+    var android = {};
+    var ios = {};
+    if (typeof unityPackages === 'object') {
+        for (var unityPackage of unityPackages) {
+            if (
+                unityPackage.variant &&
+                unityPackage.variant !== 'standard' &&
+                unityPackage.variant !== 'security'
+            ) {
+                continue;
+            }
+            if (unityPackage.platform === 'standalonewindows') {
+                if (
+                    unityPackage.performanceRating === 'None' &&
+                    pc.performanceRating
+                ) {
+                    continue;
+                }
+                pc = unityPackage;
+            } else if (unityPackage.platform === 'android') {
+                if (
+                    unityPackage.performanceRating === 'None' &&
+                    android.performanceRating
+                ) {
+                    continue;
+                }
+                android = unityPackage;
+            } else if (unityPackage.platform === 'ios') {
+                if (
+                    unityPackage.performanceRating === 'None' &&
+                    ios.performanceRating
+                ) {
+                    continue;
+                }
+                ios = unityPackage;
+            }
+        }
+    }
+    return { pc, android, ios };
+}
+
+function compareUnityVersion(unitySortNumber) {
+    if (!window.API.cachedConfig.sdkUnityVersion) {
+        console.error('No cachedConfig.sdkUnityVersion');
+        return false;
+    }
+
+    // 2022.3.6f1  2022 03 06 000
+    // 2019.4.31f1 2019 04 31 000
+    // 5.3.4p1     5    03 04 010
+    // 2019.4.31f1c1 is a thing
+    var array = window.API.cachedConfig.sdkUnityVersion.split('.');
+    if (array.length < 3) {
+        console.error('Invalid cachedConfig.sdkUnityVersion');
+        return false;
+    }
+    var currentUnityVersion = array[0];
+    currentUnityVersion += array[1].padStart(2, '0');
+    var indexFirstLetter = array[2].search(/[a-zA-Z]/);
+    if (indexFirstLetter > -1) {
+        currentUnityVersion += array[2]
+            .substr(0, indexFirstLetter)
+            .padStart(2, '0');
+        currentUnityVersion += '0';
+        var letter = array[2].substr(indexFirstLetter, 1);
+        if (letter === 'p') {
+            currentUnityVersion += '1';
+        } else {
+            // f
+            currentUnityVersion += '0';
+        }
+        currentUnityVersion += '0';
+    } else {
+        // just in case
+        currentUnityVersion += '000';
+    }
+    // just in case
+    currentUnityVersion = currentUnityVersion.replace(/\D/g, '');
+
+    if (parseInt(unitySortNumber, 10) <= parseInt(currentUnityVersion, 10)) {
+        return true;
+    }
+    return false;
+}
+
+export {
+    storeAvatarImage,
+    parseAvatarUrl,
+    getPlatformInfo,
+    compareUnityVersion
+};
