+++ src/composables/setting/constants/openSourceSoftwareLicenses.js
