--- src/composables/user/constants.js
@@ -1,126 +1,0 @@
-const userDialogGroupSortingOptions = {
-    alphabetical: {
-        name: 'dialog.user.groups.sorting.alphabetical',
-        value: 'alphabetical'
-    },
-    members: {
-        name: 'dialog.user.groups.sorting.members',
-        value: 'members'
-    },
-    inGame: {
-        name: 'dialog.user.groups.sorting.in_game',
-        value: 'inGame'
-    }
-};
-
-// vrchat to famfamfam language mappings
-const languageMappings = {
-    eng: 'us',
-    kor: 'kr',
-    rus: 'ru',
-    spa: 'es',
-    por: 'pt',
-    zho: 'cn',
-    deu: 'de',
-    jpn: 'jp',
-    fra: 'fr',
-    swe: 'se',
-    nld: 'nl',
-    pol: 'pl',
-    dan: 'dk',
-    nor: 'no',
-    ita: 'it',
-    tha: 'th',
-    fin: 'fi',
-    hun: 'hu',
-    ces: 'cz',
-    tur: 'tr',
-    ara: 'ae',
-    ron: 'ro',
-    vie: 'vn',
-    ukr: 'ua',
-    ase: 'us',
-    bfi: 'gb',
-    dse: 'nl',
-    fsl: 'fr',
-    jsl: 'jp',
-    kvk: 'kr',
-
-    mlt: 'mt',
-    ind: 'id',
-    hrv: 'hr',
-    heb: 'he',
-    afr: 'af',
-    ben: 'be',
-    bul: 'bg',
-    cmn: 'cn',
-    cym: 'cy',
-    ell: 'el',
-    est: 'et',
-    fil: 'ph',
-    gla: 'gd',
-    gle: 'ga',
-    hin: 'hi',
-    hmn: 'cn',
-    hye: 'hy',
-    isl: 'is',
-    lav: 'lv',
-    lit: 'lt',
-    ltz: 'lb',
-    mar: 'hi',
-    mkd: 'mk',
-    msa: 'my',
-    sco: 'gd',
-    slk: 'sk',
-    slv: 'sl',
-    tel: 'hi',
-    mri: 'nz',
-    wuu: 'cn',
-    yue: 'cn',
-    tws: 'cn',
-    asf: 'au',
-    nzs: 'nz',
-    gsg: 'de',
-    epo: 'eo',
-    tok: 'tok'
-};
-
-const emojiAnimationStyleUrl =
-    'https://assets.vrchat.com/www/images/emoji-previews/';
-
-const emojiAnimationStyleList = {
-    Aura: 'Preview_B2-Aura.gif',
-    Bats: 'Preview_B2-Fall_Bats.gif',
-    Bees: 'Preview_B2-Bees.gif',
-    Bounce: 'Preview_B2-Bounce.gif',
-    Cloud: 'Preview_B2-Cloud.gif',
-    Confetti: 'Preview_B2-Winter_Confetti.gif',
-    Crying: 'Preview_B2-Crying.gif',
-    Dislike: 'Preview_B2-Dislike.gif',
-    Fire: 'Preview_B2-Fire.gif',
-    Idea: 'Preview_B2-Idea.gif',
-    Lasers: 'Preview_B2-Lasers.gif',
-    Like: 'Preview_B2-Like.gif',
-    Magnet: 'Preview_B2-Magnet.gif',
-    Mistletoe: 'Preview_B2-Winter_Mistletoe.gif',
-    Money: 'Preview_B2-Money.gif',
-    Noise: 'Preview_B2-Noise.gif',
-    Orbit: 'Preview_B2-Orbit.gif',
-    Pizza: 'Preview_B2-Pizza.gif',
-    Rain: 'Preview_B2-Rain.gif',
-    Rotate: 'Preview_B2-Rotate.gif',
-    Shake: 'Preview_B2-Shake.gif',
-    Snow: 'Preview_B2-Spin.gif',
-    Snowball: 'Preview_B2-Winter_Snowball.gif',
-    Spin: 'Preview_B2-Spin.gif',
-    Splash: 'Preview_B2-SummerSplash.gif',
-    Stop: 'Preview_B2-Stop.gif',
-    ZZZ: 'Preview_B2-ZZZ.gif'
-};
-
-export {
-    userDialogGroupSortingOptions,
-    languageMappings,
-    emojiAnimationStyleUrl,
-    emojiAnimationStyleList
-};
