+++ src/composables/user/constants/emoji.js
@@ -1,0 +1,34 @@
+const emojiAnimationStyleUrl =
+    'https://assets.vrchat.com/www/images/emoji-previews/';
+
+const emojiAnimationStyleList = {
+    Aura: 'Preview_B2-Aura.gif',
+    Bats: 'Preview_B2-Fall_Bats.gif',
+    Bees: 'Preview_B2-Bees.gif',
+    Bounce: 'Preview_B2-Bounce.gif',
+    Cloud: 'Preview_B2-Cloud.gif',
+    Confetti: 'Preview_B2-Winter_Confetti.gif',
+    Crying: 'Preview_B2-Crying.gif',
+    Dislike: 'Preview_B2-Dislike.gif',
+    Fire: 'Preview_B2-Fire.gif',
+    Idea: 'Preview_B2-Idea.gif',
+    Lasers: 'Preview_B2-Lasers.gif',
+    Like: 'Preview_B2-Like.gif',
+    Magnet: 'Preview_B2-Magnet.gif',
+    Mistletoe: 'Preview_B2-Winter_Mistletoe.gif',
+    Money: 'Preview_B2-Money.gif',
+    Noise: 'Preview_B2-Noise.gif',
+    Orbit: 'Preview_B2-Orbit.gif',
+    Pizza: 'Preview_B2-Pizza.gif',
+    Rain: 'Preview_B2-Rain.gif',
+    Rotate: 'Preview_B2-Rotate.gif',
+    Shake: 'Preview_B2-Shake.gif',
+    Snow: 'Preview_B2-Spin.gif',
+    Snowball: 'Preview_B2-Winter_Snowball.gif',
+    Spin: 'Preview_B2-Spin.gif',
+    Splash: 'Preview_B2-SummerSplash.gif',
+    Stop: 'Preview_B2-Stop.gif',
+    ZZZ: 'Preview_B2-ZZZ.gif'
+};
+
+export { emojiAnimationStyleUrl, emojiAnimationStyleList };
