+++ src/composables/user/constants/language.js
@@ -1,0 +1,73 @@
+// vrchat to famfamfam language mappings
+const languageMappings = {
+    eng: 'us',
+    kor: 'kr',
+    rus: 'ru',
+    spa: 'es',
+    por: 'pt',
+    zho: 'cn',
+    deu: 'de',
+    jpn: 'jp',
+    fra: 'fr',
+    swe: 'se',
+    nld: 'nl',
+    pol: 'pl',
+    dan: 'dk',
+    nor: 'no',
+    ita: 'it',
+    tha: 'th',
+    fin: 'fi',
+    hun: 'hu',
+    ces: 'cz',
+    tur: 'tr',
+    ara: 'ae',
+    ron: 'ro',
+    vie: 'vn',
+    ukr: 'ua',
+    ase: 'us',
+    bfi: 'gb',
+    dse: 'nl',
+    fsl: 'fr',
+    jsl: 'jp',
+    kvk: 'kr',
+
+    mlt: 'mt',
+    ind: 'id',
+    hrv: 'hr',
+    heb: 'he',
+    afr: 'af',
+    ben: 'be',
+    bul: 'bg',
+    cmn: 'cn',
+    cym: 'cy',
+    ell: 'el',
+    est: 'et',
+    fil: 'ph',
+    gla: 'gd',
+    gle: 'ga',
+    hin: 'hi',
+    hmn: 'cn',
+    hye: 'hy',
+    isl: 'is',
+    lav: 'lv',
+    lit: 'lt',
+    ltz: 'lb',
+    mar: 'hi',
+    mkd: 'mk',
+    msa: 'my',
+    sco: 'gd',
+    slk: 'sk',
+    slv: 'sl',
+    tel: 'hi',
+    mri: 'nz',
+    wuu: 'cn',
+    yue: 'cn',
+    tws: 'cn',
+    asf: 'au',
+    nzs: 'nz',
+    gsg: 'de',
+    epo: 'eo',
+    tok: 'tok'
+};
+
+export { languageMappings };
