+++ src/composables/user/constants/userDialogGroupSortingOptions.js
@@ -1,0 +1,16 @@
+const userDialogGroupSortingOptions = {
+    alphabetical: {
+        name: 'dialog.user.groups.sorting.alphabetical',
+        value: 'alphabetical'
+    },
+    members: {
+        name: 'dialog.user.groups.sorting.members',
+        value: 'members'
+    },
+    inGame: {
+        name: 'dialog.user.groups.sorting.in_game',
+        value: 'inGame'
+    }
+};
+
+export { userDialogGroupSortingOptions };
