+++ src/mixins/dialogs/dialogs.pug
@@ -128,7 +128,6 @@ mixin dialogs
         :activeFriends='activeFriends'
         :onlineFriends='onlineFriends'
         :vipFriends='vipFriends'
-        :hasGroupPermission='hasGroupPermission'
         :inviteMessageTable='inviteMessageTable'
         :uploadImage='uploadImage'
         @openFolderGeneric='openFolderGeneric'
